+++ .vscode-test.mjs
@@ -2,5 +2,6 @@ import { defineConfig } from '@vscode/test-cli';
 
 export default defineConfig({
 	files: 'out/test/**/*.test.js',
+	launchArgs: ['--disable-extensions'],
 	skipExtensionDependencies: true,
 });